Fix gallery compatibility check for array values
- Extracted matchRequirement to src/utils/gallery-requirements.ts
- Added robust array matching logic (Exact Match vs List Match)
- Updated gallery routes to use new utility
- Updated tests to cover array matching cases
