+++ packages/service/src/routes/gallery.routes.ts
@@ -37,6 +37,7 @@ import {
 import type { LogRetentionService } from '../log-retention.service.js';
 import type { RateLimiter } from '../utils/rate-limiter.js';
 import type { PersistableHomenetBridgeConfig } from '../types/index.js';
+import { checkConfigRequirements, matchRequirement } from '../utils/gallery-requirements.js';
 
 export interface GalleryRoutesContext {
   configRateLimiter: RateLimiter;
@@ -224,95 +225,6 @@ export function createGalleryRoutes(ctx: GalleryRoutesContext): Router {
 
     return res.json({ compatibilityByVendorId });
   });
-
-  // Helper to check if a config matches vendor requirements
-  function checkConfigRequirements(
-    config: HomenetBridgeConfig,
-    requirements: { serial?: Record<string, unknown>; packet_defaults?: Record<string, unknown> },
-  ): boolean {
-    if (!config.serial) return false;
-
-    // Helper for flexible matching
-    const matchRequirement = (expected: unknown, actual: unknown): boolean => {
-      // 1. List matching (Array)
-      if (Array.isArray(expected)) {
-        // If actual value is in the expected array, it's a match
-        return expected.includes(actual);
-      }
-
-      // 2. Range matching (Object with min/max)
-      if (
-        typeof expected === 'object' &&
-        expected !== null &&
-        ('min' in expected || 'max' in expected)
-      ) {
-        const range = expected as { min?: number; max?: number };
-        if (typeof actual !== 'number') return false;
-
-        if (range.min !== undefined && actual < range.min) return false;
-        if (range.max !== undefined && actual > range.max) return false;
-        return true;
-      }
-
-      // 3. Exact matching (Primitive or simple object)
-      // Normalize values: treat empty arrays, null, and undefined as equivalent (empty/default)
-      const normalizeValue = (v: unknown): string | unknown => {
-        if (v === null || v === undefined) return '__EMPTY__';
-        if (Array.isArray(v) && v.length === 0) return '__EMPTY__';
-        if (Array.isArray(v)) return JSON.stringify(v);
-        return v;
-      };
-
-      return normalizeValue(expected) === normalizeValue(actual);
-    };
-
-    // Check serial settings
-    if (requirements.serial) {
-      const serialFields = ['baud_rate', 'data_bits', 'parity', 'stop_bits'];
-      for (const field of serialFields) {
-        const expected = requirements.serial[field];
-        const actual = config.serial[field as keyof typeof config.serial];
-        
-        // Skip check if requirement is not defined
-        if (expected === undefined) continue;
-        
-        // actual can be undefined in config, need to handle it.
-        // If expected is defined but actual is undefined, it's a mismatch (unless expected allows undefined/null, but usually config MUST exist)
-        if (actual === undefined && expected !== undefined) return false;
-
-        if (!matchRequirement(expected, actual)) {
-          return false;
-        }
-      }
-    }
-
-    // Check packet_defaults
-    if (requirements.packet_defaults) {
-      const packetDefaults = config.packet_defaults || {};
-      const packetFields = [
-        'rx_length',
-        'rx_checksum',
-        'tx_checksum',
-        'rx_header',
-        'tx_header',
-        'rx_footer',
-        'tx_footer',
-      ];
-
-      for (const field of packetFields) {
-        const expected = requirements.packet_defaults[field];
-        const actual = packetDefaults[field as keyof typeof packetDefaults];
-
-        if (expected !== undefined) {
-           if (!matchRequirement(expected, actual)) {
-            return false;
-          }
-        }
-      }
-    }
-
-    return true;
-  }
 
   // Evaluate discovery schemas against packet dictionary
   router.get('/api/gallery/discovery', async (req, res) => {
@@ -705,7 +617,7 @@ export function createGalleryRoutes(ctx: GalleryRoutesContext): Router {
           for (const field of serialFields) {
             const expected = vendorRequirements.serial[field];
             const actual = serialConfig[field];
-            if (expected !== undefined && actual !== undefined && expected !== actual) {
+            if (expected !== undefined && actual !== undefined && !matchRequirement(expected, actual)) {
               compatibility.compatible = false;
               compatibility.mismatches.push({
                 field: `serial.${field}`,
@@ -734,15 +646,7 @@ export function createGalleryRoutes(ctx: GalleryRoutesContext): Router {
             const actual = packetDefaults[field];
 
             if (expected !== undefined) {
-              // Normalize values: treat empty arrays, null, and undefined as equivalent (empty/default)
-              const normalizeValue = (v: unknown): string | unknown => {
-                // Empty array, null, or undefined are all considered "empty"
-                if (v === null || v === undefined) return '__EMPTY__';
-                if (Array.isArray(v) && v.length === 0) return '__EMPTY__';
-                if (Array.isArray(v)) return JSON.stringify(v);
-                return v;
-              };
-              if (normalizeValue(expected) !== normalizeValue(actual)) {
+              if (!matchRequirement(expected, actual ?? null)) {
                 compatibility.compatible = false;
                 compatibility.mismatches.push({
                   field: `packet_defaults.${field}`,
@@ -785,7 +689,7 @@ export function createGalleryRoutes(ctx: GalleryRoutesContext): Router {
       const currentConfigs = ctx.getCurrentConfigs();
       const currentConfigFiles = ctx.getCurrentConfigFiles();
 
-      // Find the config file for this portId
+      // Find the config for this portId
       let targetConfigFile: string | null = null;
       let configIndex = -1;
