+++ packages/service/test/routes/gallery-requirements-enhanced.test.ts
@@ -1,95 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { HomenetBridgeConfig } from '@rs485-homenet/core';
-
-// Re-implementing the function locally for testing purposes as it was added as a local helper
-// copy-pasted from gallery.routes.ts to ensure we test the same logic
-function checkConfigRequirements(
-  config: HomenetBridgeConfig,
-  requirements: { serial?: Record<string, unknown>; packet_defaults?: Record<string, unknown> },
-): boolean {
-  if (!config.serial) return false;
-
-  // Helper for flexible matching
-  const matchRequirement = (expected: unknown, actual: unknown): boolean => {
-    // 1. List matching (Array)
-    if (Array.isArray(expected)) {
-      // If actual value is in the expected array, it's a match
-      return expected.includes(actual);
-    }
-
-    // 2. Range matching (Object with min/max)
-    if (
-      typeof expected === 'object' &&
-      expected !== null &&
-      ('min' in expected || 'max' in expected)
-    ) {
-      const range = expected as { min?: number; max?: number };
-      if (typeof actual !== 'number') return false;
-
-      if (range.min !== undefined && actual < range.min) return false;
-      if (range.max !== undefined && actual > range.max) return false;
-      return true;
-    }
-
-    // 3. Exact matching (Primitive or simple object)
-    // Normalize values: treat empty arrays, null, and undefined as equivalent (empty/default)
-    const normalizeValue = (v: unknown): string | unknown => {
-      if (v === null || v === undefined) return '__EMPTY__';
-      if (Array.isArray(v) && v.length === 0) return '__EMPTY__';
-      if (Array.isArray(v)) return JSON.stringify(v);
-      return v;
-    };
-
-    return normalizeValue(expected) === normalizeValue(actual);
-  };
-
-  // Check serial settings
-  if (requirements.serial) {
-    const serialFields = ['baud_rate', 'data_bits', 'parity', 'stop_bits'];
-    for (const field of serialFields) {
-      const expected = requirements.serial[field];
-      const actual = config.serial[field as keyof typeof config.serial];
-      
-      // Skip check if requirement is not defined
-      if (expected === undefined) continue;
-      
-      // actual can be undefined in config, need to handle it.
-      // If expected is defined but actual is undefined, it's a mismatch (unless expected allows undefined/null, but usually config MUST exist)
-      if (actual === undefined && expected !== undefined) return false;
-
-      if (!matchRequirement(expected, actual)) {
-        return false;
-      }
-    }
-  }
-
-  // Check packet_defaults
-  if (requirements.packet_defaults) {
-    const packetDefaults = config.packet_defaults || {};
-    const packetFields = [
-      'rx_length',
-      'rx_checksum',
-      'tx_checksum',
-      'rx_header',
-      'tx_header',
-      'rx_footer',
-      'tx_footer',
-    ];
-
-    for (const field of packetFields) {
-      const expected = requirements.packet_defaults[field];
-      const actual = packetDefaults[field as keyof typeof packetDefaults];
-
-      if (expected !== undefined) {
-         if (!matchRequirement(expected, actual)) {
-          return false;
-        }
-      }
-    }
-  }
-
-  return true;
-}
+import { checkConfigRequirements } from '../../src/utils/gallery-requirements.js';
 
 describe('checkConfigRequirements Enhanced', () => {
   const baseConfig: HomenetBridgeConfig = {
@@ -186,5 +97,23 @@ describe('checkConfigRequirements Enhanced', () => {
             packet_defaults: { rx_checksum: ['none', 'sum'] }
         })
     ).toBe(true);
+  });
+
+  it('should handle rx_header exact array matching', () => {
+      // Exact match [0xaa] matches [0xaa]
+      expect(
+          checkConfigRequirements(baseConfig, {
+              packet_defaults: { rx_header: [0xaa] }
+          })
+      ).toBe(true);
+  });
+
+  it('should handle rx_header list matching', () => {
+      // List match: [[0xaa], [0xbb]] matches [0xaa]
+      expect(
+          checkConfigRequirements(baseConfig, {
+              packet_defaults: { rx_header: [[0xaa], [0xbb]] }
+          })
+      ).toBe(true);
   });
 });
