+++ packages/service/test/routes/gallery-requirements.test.ts
@@ -1,170 +1,85 @@
 import { describe, it, expect } from 'vitest';
 import { HomenetBridgeConfig } from '@rs485-homenet/core';
-
-// We need to test the checkConfigRequirements logic.
-// Since it's not exported from gallery.routes.ts and is inside the file scope,
-// we'll copy the logic here for unit testing to ensure correctness,
-// or simpler, we can extract it to a utility file if we wanted to be cleaner.
-// For now, I will create a test that mocks the matching logic to verify expected behavior.
-
-// Re-implementing the function locally for testing purposes as it was added as a local helper
-function checkConfigRequirements(
-  config: HomenetBridgeConfig,
-  requirements: { serial?: Record<string, unknown>; packet_defaults?: Record<string, unknown> },
-): boolean {
-  if (!config.serial) return false;
-
-  // Check serial settings
-  if (requirements.serial) {
-    const serialFields = ['baud_rate', 'data_bits', 'parity', 'stop_bits'];
-    for (const field of serialFields) {
-      const expected = requirements.serial[field];
-      const actual = config.serial[field as keyof typeof config.serial];
-      if (expected !== undefined && actual !== undefined && expected !== actual) {
-        return false;
-      }
-    }
-  }
-
-  // Check packet_defaults
-  if (requirements.packet_defaults) {
-    const packetDefaults = config.packet_defaults || {};
-    const packetFields = [
-      'rx_length',
-      'rx_checksum',
-      'tx_checksum',
-      'rx_header',
-      'tx_header',
-      'rx_footer',
-      'tx_footer',
-    ];
-
-    for (const field of packetFields) {
-      const expected = requirements.packet_defaults[field];
-      const actual = packetDefaults[field as keyof typeof packetDefaults];
-
-      if (expected !== undefined) {
-        const normalizeValue = (v: unknown): string | unknown => {
-          if (v === null || v === undefined) return '__EMPTY__';
-          if (Array.isArray(v) && v.length === 0) return '__EMPTY__';
-          if (Array.isArray(v)) return JSON.stringify(v);
-          return v;
-        };
-
-        if (normalizeValue(expected) !== normalizeValue(actual)) {
-          return false;
-        }
-      }
-    }
-  }
-
-  return true;
-}
+import { checkConfigRequirements } from '../../src/utils/gallery-requirements.js';
 
 describe('checkConfigRequirements', () => {
-  it('should return true when requirements match config', () => {
-    const config: HomenetBridgeConfig = {
+  const baseConfig: HomenetBridgeConfig = {
       serial: {
-        port: '/dev/ttyUSB0',
+        path: '/dev/ttyUSB0',
         baud_rate: 9600,
         data_bits: 8,
         parity: 'none',
         stop_bits: 1,
+        portId: 'default'
       },
       packet_defaults: {
         rx_header: [0xaa],
       },
-      mqtt: { broker_url: '' }, // minimal required
-    };
+  };
 
+  it('should return true when requirements match config', () => {
     const requirements = {
       serial: { baud_rate: 9600, parity: 'none' },
       packet_defaults: { rx_header: [0xaa] },
     };
 
-    expect(checkConfigRequirements(config, requirements)).toBe(true);
+    expect(checkConfigRequirements(baseConfig, requirements)).toBe(true);
   });
 
   it('should return false when baud_rate mismatches', () => {
-    const config: HomenetBridgeConfig = {
-      serial: {
-        port: '/dev/ttyUSB0',
-        baud_rate: 9600,
-        data_bits: 8,
-        parity: 'none',
-        stop_bits: 1,
-      },
-      mqtt: { broker_url: '' },
-    };
-
     const requirements = {
       serial: { baud_rate: 19200 },
     };
 
-    expect(checkConfigRequirements(config, requirements)).toBe(false);
+    expect(checkConfigRequirements(baseConfig, requirements)).toBe(false);
   });
 
   it('should return false when packet_defaults mismatch', () => {
-    const config: HomenetBridgeConfig = {
-      serial: {
-        port: '/dev/ttyUSB0',
-        baud_rate: 9600,
-        data_bits: 8,
-        parity: 'none',
-        stop_bits: 1,
-      },
-      packet_defaults: {
-        rx_header: [0xaa],
-      },
-      mqtt: { broker_url: '' },
-    };
-
     const requirements = {
       packet_defaults: { rx_header: [0xbb] },
     };
 
-    expect(checkConfigRequirements(config, requirements)).toBe(false);
+    expect(checkConfigRequirements(baseConfig, requirements)).toBe(false);
   });
 
   it('should treat null/undefined/empty array as equivalent in packet_defaults', () => {
     const config: HomenetBridgeConfig = {
       serial: {
-        port: '/dev/ttyUSB0',
+        path: '/dev/ttyUSB0',
         baud_rate: 9600,
         data_bits: 8,
         parity: 'none',
         stop_bits: 1,
+        portId: 'default'
       },
       packet_defaults: {
         rx_header: [], // empty array
       },
-      mqtt: { broker_url: '' },
     };
 
     // requirement expects null (should match empty array)
+    // Note: If expected is null, normalizeValue returns '__EMPTY__'.
+    // If actual is [], normalizeValue returns '__EMPTY__'.
+    // Match!
     const req1 = {
       packet_defaults: { rx_header: null },
     };
-    expect(checkConfigRequirements(config, req1)).toBe(true);
-
-    // requirement expects undefined (should be ignored effectively if not present in object loop, but if passed explicitly as undefined it might trigger match logic depending on loop, but here JSON.stringify handles it)
-    // Actually our logic says: if expected !== undefined.
-    // If I explicitly pass undefined in the object, the loop 'for (const field of packetFields)'
-    // will see expected as undefined and skip the check.
+    // Type casting because HomenetBridgeConfig types might be strict but runtime allows it
+    expect(checkConfigRequirements(config, req1 as any)).toBe(true);
 
     // Test explicit empty array requirement matching null config
     const config2: HomenetBridgeConfig = {
       serial: {
-        port: '/dev/ttyUSB0',
+        path: '/dev/ttyUSB0',
         baud_rate: 9600,
         data_bits: 8,
         parity: 'none',
         stop_bits: 1,
+        portId: 'default'
       },
       packet_defaults: {
         // rx_header undefined
       },
-      mqtt: { broker_url: '' },
     };
 
     const req2 = {
